refactor: rename Mock macro to AddMock
