+++ Sources/TestDRS/MockPropertyMacro.swift
@@ -5,7 +5,7 @@
 
 import Foundation
 
-/// This macro is intended for use by the `MockMacro` to provide custom accessors.
+/// This macro is intended for use by the `AddMockMacro` to provide custom accessors.
 /// It should not be applied manually to properties.
 @attached(accessor)
 public macro __MockProperty() = #externalMacro(module: "TestDRSMacros", type: "MockPropertyMacro")
